Added patches to the 3rd party library, Circle. Their handling of HID class USB devices caused it to hog the CPU while continuously polling for changes. My change delays the polling by the delay reported by the device configuration. I also inserted a data memory barrier after the end of the request which seems to increase the reliability.
Also the timer interrupt handler has been changed to move the resetting of the interrupt to the end of the
handler instead of the beginning. It looks like the timer can overshoot if there are a lot of events being
processed in the handler, causing no events to be handled until the timer counter wraps around. Moving the
resetting to the end, after events have been processed may cause a greater drift, but will prevent this
from happening.

The code sections modified by me are enabled by defining a HFH3_PATCH preprocessor macro.
This way the library can be quickly built with or without my patches and it is obvious where I have
modified the original code. (Pristine code is also maintained in the vendor branch, making it easier to
update it with newer releases if needed.)