Use an interrupt service to wait for Vertical Sync.
Since the network subsystem uses Circle's cooperative task scheduler, it is no longer possible to
wait for VSync by calling CBCMFramebuffer's WaitForVerticalSync method, as it will block all 
network processing.

Instead, we can enable a feature where the GPU firmware will generate an interrupt on the ARM 
side when a vertical sync occurs. This requires that we set a flag in the firmware config.txt
file to enable the feature.

The interrupt handle is wrapped in a class that can block a task wanting to wait for the
VSync event through the Circle task scheduler, the IRQ handler will then unblock the task
when the event arrives.