Added option to render graphics into CPU memory and then memcpy the contents to the frame buffer on Present() instead of flipping between two GPU buffers.
Added timers to measure the difference between the two.
The memcpy method is faster than the GPU flip method due to caching.
When rendering directly to GPU memory, the game code is slowed down due to
lack of memory caching. This is despite the additional memory copy of the entire screen.

Using the old method, the CPU spent over 60% of the time between each vertical sync 
rendering the next frame. With the new method, around 20% of CPU time is spent rendering 
plus 2% on copying the frame buffer.