Added initial server/client architecture.
Currently the server IP address is hardcoded to 192.168.137.80.
If a machine gets that address assigned it will act as the server, otherwise it will attempt connecting to that IP address.

The client is really thin. It will simply wait for a command buffer from the server and replay it locally.
The server sends it current command buffer after generation to the client, making the client in effect a remote mirror display for
the server.