Basic multiplayer.
There are now two player objects, one for the remotely connected player.

Split the Input class into a base class and two subclasses implementing local input and proxied input.

The client now sends a byte representing the input state to the server when it changes, 
the server has a reader task that updates the state of the proxy input object representing the remote
player when it receives data from the client.

Client and server IP addresses are still hardcoded.