Instead of sending static elements to be drawn on every frame, define background elements as a grid of non-moving images drawn every frame. Map updates are now replaced with background element updates, as the Background class now takes care of updating the map.
Updated python client to handle the new command buffer commands.

Added some handling of when incomplete command buffers arrive on the client side. Command
buffers now always start with a FrameStart command containing the lenght of the buffer
sent. The client will yield the current thread to allow for more data to arrive in case
it detects a partial command buffer.